==> rscripts/univar/01_get_data_summary.rmd <==
---
title: "Bronchocela Morphology - Data Summary"
author: "Christian Supsup"
date: "13 March 2025"
output:
  html_document:
    df_print: paged
  pdf_document: default
indent: no
---
```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(knitr)
```

**Load required packages/functions**
```{r load pkgs}
## Load packages
library(psych)

## Load our custom scripts from https://github.com/csupsup/MorphoTax
library(MorphoTax)
```
**Get data summary**
```{r data summary}
#options(scipen = 999) # turn off scientific notation

## Read data
df.raw <- read.csv("data/bronchocela_morph_13March2025.csv")
df.sub <- subset(df.raw, select = c(6,11,12:35))
names(df.sub )[1] <- "Pop"
df.sub  <- df.sub [order(df.sub $Pop),]

head(df.sub )

## Replace population label with species name
df.rep <- df.sub 
df.rep$Pop <- gsub("Luzon|Lubang|Romblon|West Visayas", "B. marmorata", df.rep$Pop)
df.rep$Pop <- gsub("Mindoro", "B. carestiai", df.rep$Pop)
df.rep$Pop <- gsub("Palawan", "B. mallarii", df.rep$Pop)
df.rep$Pop <- gsub("Mindanao", "B. philippina", df.rep$Pop)
df.rep$Pop <- gsub("Borneo", "B. cristatella s.l. (Borneo)", df.rep$Pop)
df.rep$Pop <- gsub("Sumatra", "B. cristatella s.l. (Sumatra)", df.rep$Pop)
df.rep$Pop <- gsub("Moluccas", "B. cristatella s.l. (Moluccas)", df.rep$Pop)

## Correct body measurements
df.mens <- subset(df.rep, select = c(1,3:16))
df.mens.cor <- adjust_morph(df.mens, type = "multi_pop")

## Combine corrected body measurements and meristic data
df.sp <- cbind(df.mens.cor, df.rep[17:26])

## Remove outliers using interquartile range method (lower quartile = 25%, upper quartile = 75%)
df.sp <- remove_outliers(data = df.sp, char = "SVL", grp = "Pop") # Use SVL as reference and apply per population

## Get data summary, looping through all characters
df.sp$Sex <- NULL
results <- summarize_morph(df.sp, grp = "Pop")

head(results)

## Write results
write.csv(results, "results/species_morph_summary.csv", row.names = FALSE, fileEncoding = "UTF-8")

## Get character ratios for each putative species using the custom "summarize_ratio" script
## Relative difference is calculated as |FnsH-TD|/TD

## FnsH/TD
ratio <- summarize_ratio(df.sp, char.as.num = "FnsH", char.as.den = "TD", grp = "Pop")

head(ratio)

write.csv(ratio, "results/ratio_fnsh_td.csv")

## OD/HL
ratio <- summarize_ratio(df.sp, char.as.num = "TD", char.as.den = "OD", grp = "Pop")

head(ratio)

write.csv(ratio, "results/ratio_td_od.csv")

## FbL/HbL
ratio <- summarize_ratio(df.sp, char.as.num = "FbL", char.as.den = "HbL", grp = "Pop")

head(ratio)

write.csv(ratio, "results/ratio_fbl_hbl.csv")

## Data summary using all data including Java and Sulawesi
df.raw <- read.csv("data/bronchocela_morph_all_13March2025.csv")
df.sub <- subset(df.raw, select = c(6,11,12:35))
names(df.sub )[1] <- "Pop"
df.sub  <- df.sub [order(df.sub $Pop),]

df.rep <- df.sub 
df.rep$Pop <- gsub("Luzon|Lubang|Romblon|West Visayas", "marmorata", df.rep$Pop)
df.rep$Pop <- gsub("Mindoro", "carestiai", df.rep$Pop)
df.rep$Pop <- gsub("Palawan", "mallarii", df.rep$Pop)
df.rep$Pop <- gsub("Mindanao", "philippina", df.rep$Pop)
df.rep$Pop <- gsub("Borneo", "cristatella (Borneo)", df.rep$Pop)
df.rep$Pop <- gsub("Sumatra", "cristatella (Sumatra)", df.rep$Pop)
df.rep$Pop <- gsub("Moluccas", "cristatella (Moluccas)", df.rep$Pop)
df.rep$Pop <- gsub("Java", "cristatella (Java)", df.rep$Pop)
df.rep$Pop <- gsub("Sulawesi", "celebensis", df.rep$Pop)

## Correct body measurements
df.mens <- subset(df.rep, select = c(1,3:16))
df.mens.cor <- adjust_morph(df.mens, type = "multi_pop")

## Combine corrected body measurements with meristic
df.sp <- cbind(df.mens.cor, df.rep[17:26])

## Remove outliers 
df.sp <- remove_outliers(data = df.sp, char = "SVL", grp = "Pop")

## Get data summary by looping through all characters
results <- summarize_morph(df.sp, grp = "Pop")

head(results)

## Write results
write.csv(results, "results/species_all_morph_summary.csv", row.names = FALSE, fileEncoding = "UTF-8")
```
