==> rscripts/univar/02_test_sexual_dimorph.rmd <==
---
title: "Bronchocela Morphology - Test Sexual Dimorphism"
author: "Christian Supsup"
date: "13 March 2025"
output:
  html_document:
    df_print: paged
  pdf_document: default
indent: no
---
```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(knitr)
```

**Load required packages/functions**
```{r load pkgs}
## Load packages
library(Rmisc)
library(ggplot2)

## Load our custom scripts from https://github.com/csupsup/MorphoTax
library(MorphoTax)
```

**Test sexual dimorphism using SVL**
```{r test.dimorph}
## Prepare data
df.data.raw <- read.csv("data/bronchocela_morph_13March2025.csv")
df.data.sub <- subset(df.data.raw, select = c(6,11,12:35))
names(df.data.sub )[1] <- "Pop"
df.data.sub  <- df.data.sub [order(df.data.sub $Pop),]

head(df.data.sub )

## Correct body size measurements
df.sex.data <- subset(df.data.sub, select = c(1,3:16))
df.sex.data <- adjust_morph(df.sex.data , type = "multi_pop")
df.sex.data$Sex <- df.data.sub$Sex

## Remove outliers
df.sex.data <- remove_outliers(data = df.sex.data, char = "SVL", grp = "Pop")

## Data summary
count.data <- as.data.frame(table(df.sex.data$Pop, df.sex.data$Sex))
colnames(count.data) <- c("Pop", "Sex", "Count")

mean.svl <- aggregate(SVL ~ Pop + Sex, data = df.sex.data, FUN = function(x) mean(x, na.rm = TRUE))
colnames(mean.svl)[3] <- "Mean_SVL"

sex.summary <- merge(count.data, mean.svl, by = c("Pop", "Sex"))

sex.summary <- sex.summary[order(sex.summary$Pop, sex.summary$Sex), ]

head(sex.summary)

write.csv(sex.summary, "results/sex.summary.csv")

## Filter data by sex
## Remove populations with only one sex data
## After removal, ensure that each sex has more >3 samples
## Use the custom "filter_sex" script
df.sex.data <- filter_sex(df.sex.data)

## Perform Shapiro-Wilk test by sex
results <- shapiro_sex(df.sex.data, sex = "Sex", grp = "Pop")

head(results)

## Write results to a csvl file
write.csv(results, "results/sex_shapiro_results.csv")

## Perform Levene's test by sex using SVL
results <- levene_sex(df.sex.data, sex = "Sex", char = "SVL", grp = "Pop")

head(results)

## Write results to a csv file
write.csv(results, "results/sex_levene_results.csv")

## Perform T-test by sex using SVL
results <- t_test_sex(df.sex.data, sex = "Sex", char = "SVL", grp = "Pop")

head(results)

## Write results to a csv file
write.csv(results , "results/sex_t_test_results.csv")

## Plot SVL mean with error bars
svl.sum <- summarySE(df.sex.data, measurevar="SVL", groupvars=c("Sex","Pop"))

svl.error <- ggplot(svl.sum, aes(x=Pop, y=SVL, colour=Sex, group = Sex)) + 
    theme_classic() +
    geom_errorbar(aes(ymin=SVL-se, ymax=SVL+se), width=.1) +
    geom_line(linewidth = 1) +
    geom_point(size = 3) +
    scale_color_manual(name = "",labels = c("Female", "Male"), 
                       values = c("#66B2FF", "#003D73")) +
    ylab("Snout-Vent Length") +
    xlab("") +
    scale_shape_manual(name = "", labels = c("Female", "Male"), values = c(15,16)) +
    theme(axis.text = element_text(size = 15, color = "black"), 
          axis.title = element_text(size = 15, face ="bold"), 
          legend.text = element_text(size = 15),
          plot.title = element_blank(),
          legend.position = c(0.80, 0.95),
          legend.background = element_blank(),
          axis.text.x = element_text(angle = 50, hjust = 1),
          axis.text.y = element_text(size = 12),
          axis.title.x = element_blank())

svl.error

pdf("results/svl_error.pdf", width = 5, height = 5) 
svl.error
dev.off()
```
